Use decimal four amount

+++ src/LeasingNinja.Sales/Domain/Amount.cs
@@ -5,16 +5,16 @@ namespace LeasingNinja.Sales.Domain
     [ValueObject]
     public readonly struct Amount
     {
-        public int AmountValue { get; }
+        public decimal AmountValue { get; }
         public string Currency { get; }
 
-        private Amount(int amount, string currency) : this()
+        private Amount(decimal amount, string currency) : this()
         {
             AmountValue = amount;
             Currency = currency;
         }
 
-        public static Amount Of(int amount, string currency) {
+        public static Amount Of(decimal amount, string currency) {
             //assert currency != null;
 		
             return new Amount(amount, currency);
